Move hook to uniq file, fix imports
