+++ src/hooks/dao/useGovPoolUserVotingPower.ts
@@ -1,13 +1,10 @@
 import { BigNumber } from "@ethersproject/bignumber"
-import { useCallback, useEffect, useMemo, useState } from "react"
+import { useCallback, useEffect, useState } from "react"
 
-import { isAddress, parseTransactionError } from "utils"
+import { parseTransactionError } from "utils"
 import useError from "hooks/useError"
 import { ZERO } from "consts"
-import { useMultipleContractSingleData } from "state/multicall/hooks"
 import { GovUserKeeper as GovUserKeeper_ABI } from "abi"
-import { Interface } from "@ethersproject/abi"
-import { isArray } from "lodash"
 import useContract from "hooks/useContract"
 import { GovUserKeeper } from "interfaces/typechain"
 
@@ -27,18 +24,6 @@ interface IVotingPowerResponse {
   nftIds: BigNumber[]
 }
 
-interface IUserVotingPowerMulticallResponse {
-  default: {
-    [address: string]: IVotingPowerResponse
-  }
-  micropool: {
-    [address: string]: IVotingPowerResponse
-  }
-  delegated: {
-    [address: string]: IVotingPowerResponse
-  }
-}
-
 const initialState: IVotingPowerResponse = {
   power: ZERO,
   nftPower: ZERO,
@@ -46,103 +31,6 @@ const initialState: IVotingPowerResponse = {
   ownedBalance: ZERO,
   ownedLength: ZERO,
   nftIds: [],
-}
-
-const USER_KEEPER_INTERFACE = new Interface(GovUserKeeper_ABI)
-
-export const useGovPoolVotingPowerMulticall = (
-  params: IUseGovPoolUserVotingPower[]
-): [data: IUserVotingPowerMulticallResponse, loading: boolean] => {
-  const validatedList = useMemo(
-    () =>
-      params.filter(
-        (p) => isAddress(p.address) && isAddress(p.userKeeperAddress)
-      ),
-    [params]
-  )
-
-  const validatedAddresses = useMemo(
-    () => validatedList.map((p) => p.userKeeperAddress as string),
-    [validatedList]
-  )
-
-  const validatedParams = useMemo(() => {
-    const addresses = validatedList.map((p) => p.address as string)
-    const isMicroPool = validatedList.map((p) => p.isMicroPool || false)
-    const useDelegated = validatedList.map((p) => p.useDelegated || false)
-
-    return [addresses, isMicroPool, useDelegated] as const
-  }, [validatedList])
-
-  const callResults = useMultipleContractSingleData(
-    validatedAddresses,
-    USER_KEEPER_INTERFACE,
-    "votingPower",
-    validatedParams as any
-  )
-
-  const anyLoading: boolean = useMemo(
-    () => callResults.some((callState) => callState.loading),
-    [callResults]
-  )
-
-  return useMemo(() => {
-    return [
-      validatedParams.length > 0
-        ? validatedParams[0].reduce<IUserVotingPowerMulticallResponse>(
-            (memo, account, i) => {
-              const value = callResults?.[i]?.result?.[0]
-
-              if (!value) return memo
-
-              if (!isArray(value) || !value.length) return memo
-
-              params.map((p, i) => {
-                const {
-                  power,
-                  nftPower,
-                  perNftPower,
-                  ownedBalance,
-                  ownedLength,
-                  nftIds,
-                } = value[i]
-
-                const result = {
-                  power,
-                  nftPower,
-                  perNftPower,
-                  ownedBalance,
-                  ownedLength,
-                  nftIds,
-                }
-
-                if (validatedParams[1][i]) {
-                  memo.micropool[account] = result
-                }
-
-                if (validatedParams[2][i]) {
-                  memo.delegated[account] = result
-                }
-
-                memo.default[account] = result
-              })
-
-              return memo
-            },
-            {
-              default: {},
-              micropool: {},
-              delegated: {},
-            }
-          )
-        : {
-            default: {},
-            micropool: {},
-            delegated: {},
-          },
-      anyLoading,
-    ]
-  }, [params, validatedParams, anyLoading, callResults])
 }
 
 const useGovPoolUserVotingPower = ({
